0.9.0: Rewrite syncing logic
By directly unifying server and client players' main inventories, every sync request was effectively two. This can cause problems when picking items on certain screens: the cursor stack would be replaced by the stack we wanted, but then once again to an empty stack (because the slot was cleared).
This commit rewrites the code to only sync inventories as it should be, and splits the extra slots to a different inventory category which will be synced the old way.