0.9.0: Rewrite inventory swapping
The old way of swapping inventory involves modifying the actual content of the inventory. This can cause a lot of problems such as inventory not syncing, item duping,... and complicates a lot of dependent code.

Each screen has a list of tracking slots for inventory, crafting/smelting... By modifying the inventory index of the tracking slot, the displayed inventory content can be swapped without modifying the actual inventory.